Move common UI to common
