+++ composeApp/src/commonMain/kotlin/io/music_assistant/client/ui/compose/library/LibraryFilterSheet.kt
@@ -18,6 +18,8 @@ import io.music_assistant.client.data.model.client.LibraryFilters
 import io.music_assistant.client.data.model.client.MediaType
 import io.music_assistant.client.data.model.client.stringResource
 import io.music_assistant.client.ui.compose.common.DataState
+import io.music_assistant.client.ui.compose.common.MultiSelectDialog
+import io.music_assistant.client.ui.compose.common.SelectOption
 import io.music_assistant.client.ui.compose.common.SettingsSheet
 import io.music_assistant.client.ui.compose.common.SettingsSheet.MultiChoiceChipsRow
 import io.music_assistant.client.ui.compose.common.SettingsSheet.PickerRow
